requirements.txt
Django==2.1.1
django-guardian==1.4.9
django-multiselectfield==0.1.8
mysqlclient==1.3.13
Pillow==5.2.0
pytz==2018.5
six==1.11.0
